Added functionality for adding dishes to the grocerylist

+++ myapplication/src/main/java/com/example/myapplication/Database/Repository/GroceryRepository.java
@@ -25,7 +25,7 @@ public class GroceryRepository {
         return groceryDao.getAll();
     }
 
-    public List<Grocery> getById(int id){
+    public List<Grocery> getById(long id){
         return groceryDao.getById(id);
     }
